Revamp subreddit output to show title and/or public description
Some subs define a public description; some don't.

Some subs define a "title"; some don't.

There's also a description shown in the sidebar after joining, but that
one can contain Markdown and would be a lot messier to show on IRC. The
public description is what's shown in search results.