Fallback to API if cheeky hack fails to get submission from video link
Reddit doesn't always return the expected redirection when requesting
the `/video/:ID` route. Best theory (haven't bothered to capture any raw
responses from a bot with this problem) is that some IP ranges are
blacklisted, like datacenters. I observed this issue on a RackNerd VPS.

The fix is to apply the same technique used for images: reverse-search
for submissions using the `url:` operator.

I've also quoted the value of that operator in search queries, as Reddit
says you are supposed to do that in its search docs.